backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.1.1
python-redis-lock==4.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1

# LangGraph & AI - Updated to latest compatible versions
langgraph==0.2.50
langchain==0.3.27
langchain-openai==0.3.31
langchain-community==0.3.27
langchain-core==0.3.74
openai==1.51.2

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
rich==13.8.1
typer==0.12.5

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
faker==30.3.0

# Development
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1

# Monitoring
prometheus-client==0.21.0